feat: create project, add and list task items
Implement simple commands:
- init: Initialise project
- add:  Add task item to project
- list: List all task items.

Items are written and read to items file with a plain text protocol
which is written to and read from with the dir_ family of functions.

Define item data structure with a single char * field for name, the
expectation is that all tasks will simply be specified with a name and
printed/listed accordingly.