feat: set defaults for subcommands
Subcommands now allow a code to be used as the first argument, without
needing to specify the corresponding flag (i.e. -c).

NOTE: In the case of add, there is currently a potential for conflicts
with *real* item names that happen to also be in a valid item code form
-- such as an item named "item" -- however, this is deemed to be very
rare a case, and the use of `-n`/`--name` can still be used in such cases.